+++ packages/windie-sdk-js/cjs/index.js
@@ -52,6 +52,8 @@ __exportStar(require("./transport/BackendSocketFactory.js"), exports);
 var WindieBackendSocketFactory_js_1 = require("./transport/WindieBackendSocketFactory.js");
 Object.defineProperty(exports, "createWindieSdkBackendSocket", { enumerable: true, get: function () { return WindieBackendSocketFactory_js_1.createWindieSdkBackendSocket; } });
 __exportStar(require("./transport/HostedBackendHttpClient.js"), exports);
+var WindieHostedBackendHttpClient_js_1 = require("./transport/WindieHostedBackendHttpClient.js");
+Object.defineProperty(exports, "WindieSdkClient", { enumerable: true, get: function () { return WindieHostedBackendHttpClient_js_1.WindieSdkClient; } });
 __exportStar(require("./transport/ManagedBackendSession.js"), exports);
 __exportStar(require("./transport/ManagedAgentSession.js"), exports);
 var ManagedWindieAgentSession_js_1 = require("./transport/ManagedWindieAgentSession.js");