+++ packages/windie-sdk-js/cjs/transport/HostedBackendHttpClient.js
@@ -3,7 +3,7 @@
  * Implements the hosted backend http client integration for the TypeScript SDK runtime.
  */
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.WindieSdkClient = exports.AgentHostedBackendClient = void 0;
+exports.AgentHostedBackendClient = void 0;
 function resolveFetchImplementation(fetchImpl) {
     if (fetchImpl) {
         return fetchImpl;
@@ -334,4 +334,3 @@ class AgentHostedBackendClient {
     }
 }
 exports.AgentHostedBackendClient = AgentHostedBackendClient;
-exports.WindieSdkClient = AgentHostedBackendClient;